Develope

+++ src/app/controllers/course.controller.js
@@ -1,37 +1,9 @@
 const httpStatus = require('http-status-codes');
 const log = require('../services/log.service');
 const service = require('../services/course.service');
+const profileService = require('../services/profile.service');
 
 const { StatusCodes } = httpStatus;
-
-const create = async (req, res) => {
-  try {
-    const { course } = req.body;
-
-    if (!course.name) {
-      res
-        .status(StatusCodes.BAD_REQUEST)
-        .json({ error: 'O nome precisa ser preenchido' });
-    }
-
-    log.info(`Inicializando criação do course: ${course.name}`);
-    const newCourse = await service.create(course);
-
-    log.info(`Buscando curso por id = ${newCourse.id}`);
-    const newCourseInfo = await service.getById(newCourse.id);
-
-    log.info('Finalizado a criação do course.');
-    return res.status(StatusCodes.CREATED).json(newCourseInfo);
-  } catch (error) {
-    const errorMsg = 'Erro ao criar course';
-
-    log.error(errorMsg, 'app/controllers/course.controller.js', error.message);
-
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .json({ error: `${errorMsg} ${error.message}` });
-  }
-};
 
 const getById = async (req, res) => {
   try {
@@ -81,10 +53,97 @@ const getAll = async (req, res) => {
   }
 };
 
+// const getMe = async (req, res) => {
+//   try {
+//     const { user } = req;
+//     const { query } = req;
+
+//     log.info(`Iniciando listagem dos cursos, page: ${query.page}`);
+
+//     const courses = await service.getMe(user, query);
+
+//     log.info('Busca finalizada com sucesso');
+//     return res.status(StatusCodes.OK).json(courses);
+//   } catch (error) {
+//     const errorMsg = 'Erro ao buscar curso';
+
+//     log.error(errorMsg, 'app/controllers/course.controller.js', error.message);
+
+//     return res
+//       .status(StatusCodes.INTERNAL_SERVER_ERROR)
+//       .json({ error: `${errorMsg} ${error.message}` });
+//   }
+// };
+
+const getMe = async (req, res) => {
+  try {
+    const { user } = req;
+
+    log.info('Buscando curso atual:');
+    const profile = await profileService.getById(user.id);
+
+    if (!profile.currentCourseId) {
+      return res
+        .status(StatusCodes.BAD_REQUEST)
+        .json({ error: 'O usuario não tem um curso atual' });
+    }
+
+    const course = await service.getById(profile.currentCourseId);
+
+    log.info('Busca finalizada com sucesso');
+    return res.status(StatusCodes.OK).json(course);
+  } catch (error) {
+    const errorMsg = 'Erro ao buscar curso';
+
+    log.error(errorMsg, 'app/controllers/course.controller.js', error.message);
+
+    return res
+      .status(StatusCodes.INTERNAL_SERVER_ERROR)
+      .json({ error: `${errorMsg} ${error.message}` });
+  }
+};
+
+const create = async (req, res) => {
+  try {
+    const { user } = req;
+    const { newCourse } = req.body;
+
+    if (!newCourse.name) {
+      return res
+        .status(StatusCodes.BAD_REQUEST)
+        .json({ error: 'O nome precisa ser preenchido' });
+    }
+
+    const course = { ...newCourse, userId: user.id };
+
+    log.info(`Inicializando criação do course: ${course.name}`);
+    const createdCourse = await service.create(course);
+
+    log.info(`Buscando curso por id = ${createdCourse.id}`);
+    const includeDiscipline = 'false';
+    const newCourseInfo = await service.getById(
+      createdCourse.id,
+      includeDiscipline,
+    );
+
+    log.info('Finalizado a criação do course.');
+    return res.status(StatusCodes.CREATED).json(newCourseInfo);
+  } catch (error) {
+    const errorMsg = 'Erro ao criar course';
+
+    log.error(errorMsg, 'app/controllers/course.controller.js', error.message);
+
+    return res
+      .status(StatusCodes.INTERNAL_SERVER_ERROR)
+      .json({ error: `${errorMsg} ${error.message}` });
+  }
+};
+
 const edit = async (req, res) => {
   try {
     const { id } = req.params;
     const { course } = req.body;
+    const { user } = req;
 
     log.info(`Iniciando atualização do curso. courseId = ${id}`);
     log.info('Verificando se curso existe');
@@ -92,9 +151,15 @@ const edit = async (req, res) => {
     const existedCourse = await service.getJustCourseById(id);
 
     if (!existedCourse) {
-      res
+      return res
         .status(StatusCodes.BAD_REQUEST)
         .json({ error: 'curso não encontrado' });
+    }
+
+    if (existedCourse.userId !== user.id) {
+      return res
+        .status(StatusCodes.BAD_REQUEST)
+        .json({ error: 'Você não pode alterar esse curso' });
     }
 
     if (course?.name) {
@@ -115,7 +180,8 @@ const edit = async (req, res) => {
     }
 
     log.info('Buscando dados atualizados do curso');
-    const courseInfo = await service.getById(id);
+    const includeDiscipline = 'false';
+    const courseInfo = await service.getById(id, includeDiscipline);
 
     log.info('Finalizando atualização');
     return res.status(StatusCodes.OK).json(courseInfo);
@@ -133,6 +199,7 @@ const edit = async (req, res) => {
 const deleteCourse = async (req, res) => {
   try {
     const { id } = req.params;
+    const { user } = req;
 
     log.info(`Iniciando remoção de curso. courseId = ${id}`);
 
@@ -142,6 +209,12 @@ const deleteCourse = async (req, res) => {
       return res
         .status(StatusCodes.NOT_FOUND)
         .json({ error: 'curso não encontrado' });
+    }
+
+    if (course.userId !== user.id) {
+      return res
+        .status(StatusCodes.NOT_FOUND)
+        .json({ error: 'Você nao pode deletar esse curso' });
     }
 
     await service.deleteCourse(course);
@@ -160,9 +233,10 @@ const deleteCourse = async (req, res) => {
 };
 
 module.exports = {
-  create,
   getById,
   getAll,
+  getMe,
+  create,
   edit,
   deleteCourse,
 };
